chore: code & support itemsConfig
