+++ src/App.tsx
@@ -1,10 +1,9 @@
 import { Col, Row } from 'antd';
-import Renderer from 'components/Renderer';
-import Sidebar from 'components/Sidebar';
-import { DEFAULT_EDGE_TYPES, DEFAULT_NODE_TYPES } from 'configs';
+import Renderer, { RendererProps } from 'components/Renderer';
+import Sidebar, { SidebarProps } from 'components/Sidebar';
+import { DEFAULT_EDGE_TYPES, DEFAULT_ITEMS_CONFIG, DEFAULT_NODE_TYPES } from 'configs';
 import React, { FC } from 'react';
-import { ReactFlowProps } from 'react-flow-renderer';
-import { assign } from 'utils';
+import { assign, assignArray } from 'utils';
 
 import styles from './App.module.less';
 
@@ -12,14 +11,14 @@ console.log(import.meta.env);
 
 const showSidebar = sessionStorage.getItem('show-side-bar');
 
-interface AppProps extends ReactFlowProps {}
+interface AppProps extends SidebarProps, RendererProps {}
 
-const App: FC<AppProps> = ({ nodeTypes, edgeTypes, ...rest }) => {
+const App: FC<AppProps> = ({ nodeTypes, edgeTypes, itemsConfig, ...rest }) => {
   return (
     <Row className={`${styles.app} app`}>
       {showSidebar && (
         <Col flex="300px">
-          <Sidebar />
+          <Sidebar itemsConfig={assignArray(DEFAULT_ITEMS_CONFIG, itemsConfig)} />
         </Col>
       )}
       <Col flex="auto">
