+++ src/components/Sidebar.tsx
@@ -1,5 +1,4 @@
-import { itemsConfig } from 'configs';
-import React from 'react';
+import React, { FC } from 'react';
 import { stringifyJSON } from 'utils';
 
 import styles from './Sidebar.module.less';
@@ -9,10 +8,19 @@ const onDragStart = (event: React.DragEvent<HTMLLIElement>, allData: string) => 
   event.dataTransfer.effectAllowed = 'move';
 };
 
-const Sidebar = () => {
+export interface SidebarProps {
+  itemsConfig?: {
+    id: number;
+    nodeType: string;
+    label: string;
+    data: Record<string, any>;
+  }[];
+}
+
+const Sidebar: FC<SidebarProps> = ({ itemsConfig }) => {
   return (
     <ul className={styles.sidebar}>
-      {itemsConfig.map(({ id, nodeType, label, data }) => (
+      {itemsConfig?.map(({ id, nodeType, label, data }) => (
         <li
           key={id}
           className={styles.sidebarItem}
