+++ src/configs/constant.ts
@@ -1,3 +1,5 @@
 export const DEFAULT_NODE_TYPES = {};
 
 export const DEFAULT_EDGE_TYPES = {};
+
+export const DEFAULT_ITEMS_CONFIG = [];
