+++ src/configs/index.ts
@@ -1,2 +1,1 @@
 export * from './constant';
-export * from './sidebar';
