+++ src/utils/assign.ts
@@ -1,3 +1,7 @@
-export const assign = (target = {}, source = {}, defaultValue = {}) => {
-  return Object.assign({}, target, source || defaultValue);
+export const assign = (target: object, source: object = {}) => {
+  return Object.assign({}, target, source);
+};
+
+export const assignArray = (target: any[], source: any[] = []) => {
+  return [...target, ...source];
 };
